Add raw string passing to fix title issues
The problem is simple enough: the extra backslashes or percents that
__lp_escape() adds show up in the title, as the title is changed in the
previous commit to no longer be printed as part of PS1. The fix is not
so simple.

The issue is the "old" Liquidprompt v1 templates/themes. Needing to
support the old $LP_BATT and similar is making all of this very
complicated. But I get why supporting it is needed: not only do we not
want to break any old user configs, but for anyone that simply wants to
put a specific dev env in front of another, they have no reason to build
a whole theme for that.

This patch moves the data sanitizing with __lp_escape() to the theme
functions instead, and from there returns two copies of the strings: one
"raw", with no color or shell character escaping, and one with all that
for saving to PS1.

This will probably run a bit faster too, since there are less
__lp_escape() calls than before. By delaying the call as long as
possible, some calls are now run as one call, and some can be skipped if
the returned string is never used.

There are a few large issues standing:

To prevent LP_PS1_PREFIX, LP_PS1_POSTFIX, and LP_MARK_PREFIX that have
color from messing up the title, the whole title string is still passed
into lp_formatted_title() to strip color out. This is less than ideal,
but faster than treating those 3 variables by themselves.

__lp_escape() was changed to a private ("__") function in v2.0, but this
change moves its indented usage to a theme ("_") level. Not only is the
name mis-match frustrating, but its usage is anoying as well. Making
theme authors deal with this function, as well as with two copies of all
strings is more than doubling the work to make a theme.

Missing yet in this commit is documentation with these changes, and
changes to the powerline themes.